Create `MessageManager` to manage message related function
