+++ bambooai/bambooai.py
@@ -9,6 +9,7 @@ warnings.filterwarnings('ignore')
 
 from bambooai import code_executor, models, template_formatting, qa_retrieval, log_manager, output_manager, web_output_manager, storage_manager, utils, executor_client
 from bambooai.messages import reg_ex, tools_definition
+from bambooai.messages.message_manager import MessageManager
 from bambooai.messages.prompts import PromptManager
 
 class BambooAI:
@@ -43,13 +44,6 @@ class BambooAI:
 
         # Web search mode
         self.search_mode = os.getenv('WEB_SEARCH_MODE', 'google_ai')
-
-        # Storage
-        try:
-            self.interaction_store = storage_manager.SimpleInteractionStore()
-        except storage_manager.StorageError as e:
-            self.output_manager.print_wrapper(f"Warning: Failed to initialize interaction store: {e}")
-            self.interaction_store = None
 
         # Check if the OPENAI_API_KEY environment variable is set
         if not os.getenv('OPENAI_API_KEY'):
@@ -97,11 +91,8 @@ class BambooAI:
                 vector_db = False
 
         self.MAX_ERROR_CORRECTIONS = 5
-        # Set the maximum number of question/answer pairs to be kept in the conversation memmory
-        self.MAX_CONVERSATIONS = (max_conversations*2) - 1
         
         # Dataframe
-
         # Check if the dataframe is provided and generate a unique ID if not provided
         if df is not None and not df_id:
             df_id = str(uuid.uuid4())
@@ -117,21 +108,6 @@ class BambooAI:
 
         # User intent
         self.user_intent = None
-        
-        # Results of the code execution
-        self.code_exec_results = None
-
-        # QA Pairs (a list of dictionaries containing "question/results of the code execution" pairs)
-        self.qa_pairs = []
-        
-        # Tasks (a list of task)
-        self.tasks = []
-
-        # Last generated code
-        self.last_code = None
-
-        # Last generated plan
-        self.last_plan = None
 
         # Set the vector_db mode. This mode is True when you want the model to rank the generated code, and store the results above threshold in a vector database.
         self.vector_db = vector_db
@@ -175,15 +151,9 @@ class BambooAI:
         # Prompt manager
         self.prompts = PromptManager(custom_prompt_file_path=custom_prompt_file)
 
-
-        # Messages lists
-        self.pre_eval_messages = [{"role": "system", "content": self.prompts.expert_selector_system}]
-        self.select_analyst_messages = [{"role": "system", "content": self.prompts.analyst_selector_system}]
-        self.eval_messages = [{"role": "system", "content": self.prompts.planner_system}]
-        self.code_messages = [{"role": "system", "content": self.prompts.code_generator_system_df}]
-        self.df_inspector_messages = [{"role": "system", "content": self.prompts.dataframe_inspector_system}]
-        self.plan_review_messages = None
-        self.insight_messages = None
+        self.message_manager = MessageManager(output_manager=self.output_manager,
+                                               prompts=self.prompts,
+                                               model_dict=self.model_dict)
 
         # QA Retrieval
         self.similarity_threshold = 0.80
@@ -215,18 +185,8 @@ class BambooAI:
     ######################
 
     def reset_messages_and_logs(self):
-        self.pre_eval_messages = [{"role": "system", "content": self.prompts.expert_selector_system}]
-        self.select_analyst_messages = [{"role": "system", "content": self.prompts.analyst_selector_system}]
-        self.eval_messages = [{"role": "system", "content": self.prompts.planner_system}]
-        self.code_messages = [{"role": "system", "content": self.prompts.code_generator_system_df}]
-        self.df_inspector_messages = [{"role": "system", "content": self.prompts.dataframe_inspector_system}]
-        self.code_exec_results = None
-
+        self.message_manager.reset_messages()
         self.log_and_call_manager.clear_run_logs()
-
-    def reset_non_cumul_messages(self):
-        self.plan_review_messages = None
-        self.insight_messages = None
 
     def reset_retrieved_data(self):
         self.retrieved_data_model = None
@@ -234,242 +194,6 @@ class BambooAI:
         self.retrieved_code = None
         self.retrieved_similarity_score = None
         self.retrieved_rank = None
-
-    def messages_maintenace(self, messages):
-        # Remove tool_calls messages from the messages list
-        for i in range(len(messages) - 1, -1, -1):  # Start from the last item to index 0
-            msg = messages[i]
-            if "tool_calls" in msg or msg.get("role") == "tool":
-                messages.pop(i)
-        # Remove the oldest conversation from the messages list
-        if len(messages) > self.MAX_CONVERSATIONS:
-            messages.pop(1)
-            messages.pop(1)
-            self.output_manager.display_system_messages("Truncating messages")
-    
-    def append_qa_pair(self, question, results):
-        # Define the custom operation identifier strings
-        custom_identifiers = [
-            "User requested to run the code to do a custom analysis of the activity with ID:",
-            "User manually edited your code, and requested to run it, and return the result."
-        ]
-        
-        # Remove all existing custom operation entries
-        self.qa_pairs = [
-            pair for pair in self.qa_pairs 
-            if not any(identifier in pair["task"] for identifier in custom_identifiers)
-        ]
-        
-        # Append the new QA pair
-        self.qa_pairs.append({"task": question, "result": results})
-
-    def format_qa_pairs(self, qa_pairs, max_qa_pairs=8):
-        # Format QA pairs for prompts
-        if not qa_pairs:
-            return "No previous analyses."
-        
-        # Trim qa_pairs first if it exceeds max_qa_pairs
-        if len(qa_pairs) > max_qa_pairs:
-            qa_pairs = qa_pairs[-max_qa_pairs:]  # Keep only the most recent pairs
-        
-        formatted_str = ["Previous Analyses:"]
-        
-        for i, pair in enumerate(qa_pairs, 1):
-            # Add question with minimal formatting
-            formatted_str.append(f"\n{i}. Task: {pair['task']}")
-            
-            # Format answer with minimal separators and line preservation
-            answer_lines = [line for line in pair['result'].split('\n') if line.strip()]
-            formatted_str.append('Result:\n' + '\n'.join(answer_lines))
-            
-            # Add minimal separator if not the last pair
-            if i < len(qa_pairs):
-                formatted_str.append("-" * 5)
-         
-        return '\n'.join(formatted_str)
-    
-    def format_tasks(self, tasks):
-        # Format tasks for prompts
-        if not tasks:
-            return "No previous tasks."
-        
-        formatted_str = ["Tasks:"]
-        
-        for i, task in enumerate(tasks, 1):
-            # Add task with minimal formatting
-            formatted_str.append(f"\n{i}. {task}")
-            
-            # Add minimal separator if not the last task
-            if i < len(tasks):
-                formatted_str.append("-" * 5)
-        
-        return '\n'.join(formatted_str)
-
-    def messages_content_maintenance(self, agent, messages):
-        if agent == 'Dataframe Inspector':
-            for msg in messages:
-                if msg.get('role') == 'user':
-                    content = msg.get('content', '')
-                    # Check if content is a list (image message format)
-                    if isinstance(content, list):
-                        # Extract only the text content from the message
-                        text_content = next((item.get('text', '') for item in content if item.get('type') == 'text'), '')
-                        # Replace the content with just the text and process it
-                        msg['content'] = reg_ex._remove_all_except_task_ontology_text(text_content)
-                    else:
-                        # Process regular text message
-                        msg['content'] = reg_ex._remove_all_except_task_ontology_text(content)
-        # Planner messages content maintenance
-        if agent == 'Planner' or agent == 'Theorist':
-            for msg in messages:
-                if msg.get('role') == 'user':
-                    content = msg.get('content', '')
-                    # Check if content is a list (image message format)
-                    if isinstance(content, list):
-                        # Extract only the text content from the message
-                        text_content = next((item.get('text', '') for item in content if item.get('type') == 'text'), '')
-                        # Replace the content with just the text and process it
-                        msg['content'] = reg_ex._remove_all_except_task_xml(text_content)
-                    else:
-                        # Process regular text message
-                        msg['content'] = reg_ex._remove_all_except_task_xml(content)
-
-        if agent == 'Code Executor':
-            for msg in messages:
-                if msg.get('role') == 'user':
-                    content = msg.get('content', '')
-                    # Check if content is a list (image message format)
-                    if isinstance(content, list):
-                        # Extract only the text content from the message
-                        text_content = next((item.get('text', '') for item in content if item.get('type') == 'text'), '')
-                        # Replace the content with just the text and process it
-                        if self.model_dict[models.get_model_name('Code Generator')[0]]['templ_formating'] == 'xml':
-                            msg['content'] = reg_ex._remove_all_except_task_xml(text_content)
-                        else:  # templ_formating is 'text'
-                            msg['content'] = reg_ex._remove_all_except_task_text(text_content)
-                    else:
-                        # Process regular text message
-                        if self.model_dict[models.get_model_name('Code Generator')[0]]['templ_formating'] == 'xml':
-                            msg['content'] = reg_ex._remove_all_except_task_xml(content)
-                        else:  # templ_formating is 'text'
-                            msg['content'] = reg_ex._remove_all_except_task_text(content)
-
-    def format_image_message(self, agent, message_content, image, provider, model):
-        # Format the message content with an image. 
-        # This replaces the last message content with the image and message content.
-        
-        if agent == 'Code Generator' or agent == 'Planner':
-            message_content = message_content
-        else:
-            message_content = self.prompts.plot_query.format(message_content)
-        
-        if model in self.multimodal_models:
-            if provider == 'anthropic':
-                return {
-                    "role": "user",
-                    "content": [
-                        {
-                            "type": "image",
-                            "source": {
-                                "type": "base64",
-                                "media_type": "image/png",
-                                "data": image,
-                            },
-                        },
-                        {
-                            "type": "text",
-                            "text": message_content
-                        }
-                    ]
-                }
-            elif provider == 'openai':
-                return {
-                    "role": "user",
-                    "content": [
-                        {
-                            "type": "image_url",
-                            "image_url": {
-                                "url": f"data:image/png;base64,{image}"
-                            }
-                        },
-                        {
-                            "type": "text",
-                            "text": message_content
-                        }
-                    ]
-                }
-            elif provider == 'gemini':
-                # For Gemini, we'll return a format that can be converted to types.ContentDict later
-                return {
-                    "role": "user",
-                    "content": [
-                        {
-                            "type": "image_base64",
-                            "data": image,
-                            "mime_type": "image/png"
-                        },
-                        {
-                            "type": "text",
-                            "text": message_content
-                        }
-                    ]
-                }
-        else:
-            return {
-                "role": "user",
-                "content": message_content
-            }
-
-    def store_interaction(self, google_search_results=None, executed_code=None, code_exec_results=None, plot_jsons=None, qa_pairs=None, tasks=None):
-        # Store the interaction if store is available
-        if hasattr(self, 'interaction_store') and self.interaction_store is not None:
-            try:
-                messages = {
-                    'pre_eval_messages': self.pre_eval_messages,
-                    'select_analyst_messages': self.select_analyst_messages,
-                    'eval_messages': self.eval_messages,
-                    'df_inspector_messages': self.df_inspector_messages,
-                    'code_messages': self.code_messages,
-                    'plan_review_messages': self.plan_review_messages,
-                    'insight_messages': self.insight_messages
-                }
-                
-                # Collect all search triplets maintaining query-links relationship
-                search_results = {
-                    'search': {
-                        'searches': [] # List of search triplets
-                    }
-                }
-                
-                # Add search triplets if they exist
-                if google_search_results:
-                    for triplet in google_search_results:
-                        search_results['search']['searches'].append({
-                            'query': triplet['query'],
-                            'result': triplet['result'],
-                            'links': triplet['links']
-                        })
-                
-                # Combine search results with code execution results
-                tool_results = {
-                    'search': search_results['search'] if search_results['search']['searches'] else None,
-                    'code_exec': {
-                        'executed_code': executed_code,
-                        'code_exec_results': code_exec_results,
-                        'plot_jsons': plot_jsons if plot_jsons else [],
-                        'qa_pairs': qa_pairs if qa_pairs else [],
-                        'tasks': tasks if tasks else []
-                    }
-                }
-                
-                self.interaction_store.store_interaction(
-                    thread_id=str(self.thread_id),
-                    chain_id=str(self.chain_id),
-                    messages=messages,
-                    tool_results=tool_results
-                )
-            except storage_manager.StorageError as e:
-                self.output_manager.print_wrapper(f"Warning: Failed to store interaction: {e}")
 
     ######################
     ### Eval Functions ###
@@ -555,7 +279,7 @@ class BambooAI:
         # Format messages if image is provided
         if image is not None:
             last_message = eval_messages[-1]
-            formatted_message = self.format_image_message(
+            formatted_message = self.message_manager.format_image_message(
                 agent=agent,
                 message_content=last_message["content"],
                 image=image,
@@ -600,21 +324,21 @@ class BambooAI:
             template = self.prompts.plot_query_routing.format(question)
         else:
             template = self.prompts.expert_selector_user.format(question)
-        self.pre_eval_messages.append({"role": "user", "content": template})
-        select_expert_llm_response, expert,requires_dataset,confidence  = self.select_expert(self.pre_eval_messages) 
-        self.pre_eval_messages.append({"role": "assistant", "content": select_expert_llm_response})
+        self.message_manager.pre_eval_messages.append({"role": "user", "content": template})
+        select_expert_llm_response, expert,requires_dataset,confidence  = self.select_expert(self.message_manager.pre_eval_messages) 
+        self.message_manager.pre_eval_messages.append({"role": "assistant", "content": select_expert_llm_response})
 
         ######## Refine Expert Selection, and Formulate the task for the expert ###########
         if expert == 'Data Analyst':
-            self.select_analyst_messages.append({"role": "user", "content": self.prompts.analyst_selector_user.format(self.format_tasks(self.tasks),
+            self.message_manager.select_analyst_messages.append({"role": "user", "content": self.prompts.analyst_selector_user.format(self.message_manager.format_tasks(),
                                                                                                               None if self.df_id is None else df_columns,
                                                                                                               aux_datasets_columns, 
                                                                                                               question)
                                                                                                               })
             
-            select_analyst_llm_response, analyst, query_unknown, query_condition, data_descr, intent_breakdown = self.select_analyst(self.select_analyst_messages)
+            select_analyst_llm_response, analyst, query_unknown, query_condition, data_descr, intent_breakdown = self.select_analyst(self.message_manager.select_analyst_messages)
             self.user_intent = intent_breakdown
-            self.select_analyst_messages.append({"role": "assistant", "content": select_analyst_llm_response})
+            self.message_manager.select_analyst_messages.append({"role": "assistant", "content": select_analyst_llm_response})
 
             self.output_manager.display_results(chain_id=self.chain_id, query={"expert":analyst, 
                                                                                "original_question": question, 
@@ -638,14 +362,14 @@ class BambooAI:
                     query = intent_breakdown
                     data_model, df_inspector_messages = utils.inspect_dataframe(df=self.df, prompt_manager=self.prompts, log_and_call_manager=self.log_and_call_manager, output_manager=self.output_manager, 
                                                                                      chain_id=self.chain_id, query=query, execution_mode=self.execution_mode, df_ontology=self.df_ontology, 
-                                                                                     df_id=self.df_id, aux_file_paths=self.auxiliary_datasets, executor_client=self.api_client, messages=self.df_inspector_messages)
+                                                                                     df_id=self.df_id, aux_file_paths=self.auxiliary_datasets, executor_client=self.api_client, messages=self.message_manager.df_inspector_messages)
                     if df_inspector_messages:
-                        self.df_inspector_messages = df_inspector_messages
+                        self.message_manager.df_inspector_messages = df_inspector_messages
                     self.data_model = reg_ex._extract_data_model(data_model)
                     dataframe_head = utils.inspect_dataframe(df=self.df, execution_mode=self.execution_mode, df_id=self.df_id, executor_client=self.api_client)
                     data_model_vis = utils.generate_model_graph(self.data_model)
-                    self.messages_maintenace(self.df_inspector_messages)
-                    self.messages_content_maintenance("Dataframe Inspector", self.df_inspector_messages)
+                    self.message_manager.messages_maintenace(self.message_manager.df_inspector_messages)
+                    self.message_manager.messages_content_maintenance("Dataframe Inspector", self.message_manager.df_inspector_messages, self.model_dict[models.get_model_name('Code Generator')[0]]['templ_formating'])
                     # Create a dictionary containing both the visualization and the YAML data
                     data_model_web = {
                         'visualization': data_model_vis,
@@ -660,16 +384,16 @@ class BambooAI:
                     self.data_model = None
 
                 if models.get_model_name("Planner")[0] not in self.reasoning_models:
-                    self.eval_messages.append({"role": "user", "content": self.prompts.planner_user_df.format(utils.get_readable_date(), 
-                                                                                                      self.format_qa_pairs(self.qa_pairs), 
+                    self.message_manager.eval_messages.append({"role": "user", "content": self.prompts.planner_user_df.format(utils.get_readable_date(), 
+                                                                                                      self.message_manager.format_qa_pairs(), 
                                                                                                       intent_breakdown, 
                                                                                                       None if self.df_id is None else dataframe_head, 
                                                                                                       utils.aux_datasets_to_string(file_paths=self.auxiliary_datasets,execution_mode=self.execution_mode,executor_client=self.api_client),
                                                                                                       f"```yaml\n{self.data_model}\n```", example_plan)
                                                                                                       })
                 else:
-                    self.eval_messages.append({"role": "user", "content": self.prompts.planner_user_df_reasoning.format(utils.get_readable_date(), 
-                                                                                                                self.format_qa_pairs(self.qa_pairs), 
+                    self.message_manager.eval_messages.append({"role": "user", "content": self.prompts.planner_user_df_reasoning.format(utils.get_readable_date(), 
+                                                                                                                self.message_manager.format_qa_pairs(), 
                                                                                                                 intent_breakdown, 
                                                                                                                 None if self.df_id is None else dataframe_head,
                                                                                                                 utils.aux_datasets_to_string(file_paths=self.auxiliary_datasets,execution_mode=self.execution_mode,executor_client=self.api_client),
@@ -677,60 +401,60 @@ class BambooAI:
                                                                                                                 })
 
                 if not self.planning: 
-                    self.eval_messages.append({"role": "assistant", "content": "No content, as planning was disabled for this task"}) # We just add a dummy assistant message if planning is disabled, to maintain the conversation messages structure
+                    self.message_manager.eval_messages.append({"role": "assistant", "content": "No content, as planning was disabled for this task"}) # We just add a dummy assistant message if planning is disabled, to maintain the conversation messages structure
 
                 # Replace first dict in messages with a new system task. This is to distinguish between the two types of analysts
-                self.code_messages[0] = {"role": "system", "content": self.prompts.code_generator_system_df}
+                self.message_manager.code_messages[0] = {"role": "system", "content": self.prompts.code_generator_system_df}
                 
             elif analyst == 'Data Analyst Generic':
                 if models.get_model_name("Planner")[0] not in self.reasoning_models:
-                    self.eval_messages.append({"role": "user", "content": self.prompts.planner_user_gen.format(utils.get_readable_date(), 
-                                                                                                       self.format_qa_pairs(self.qa_pairs), 
+                    self.message_manager.eval_messages.append({"role": "user", "content": self.prompts.planner_user_gen.format(utils.get_readable_date(), 
+                                                                                                       self.message_manager.format_qa_pairs(), 
                                                                                                        intent_breakdown, example_plan)
                                                                                                        })
                 else:
-                    self.eval_messages.append({"role": "user", "content": self.prompts.planner_user_gen_reasoning.format(utils.get_readable_date(), 
-                                                                                                                 self.format_qa_pairs(self.qa_pairs), 
+                    self.message_manager.eval_messages.append({"role": "user", "content": self.prompts.planner_user_gen_reasoning.format(utils.get_readable_date(), 
+                                                                                                                 self.message_manager.format_qa_pairs(), 
                                                                                                                  intent_breakdown)
                                                                                                                  })
 
                 if not self.planning:
-                    self.eval_messages.append({"role": "assistant", "content": "No content, as planning was disabled for this task"}) # We just add a dummy assistant message if planning is disabled, to maintain the conversation messages structure
+                    self.message_manager.eval_messages.append({"role": "assistant", "content": "No content, as planning was disabled for this task"}) # We just add a dummy assistant message if planning is disabled, to maintain the conversation messages structure
 
                 # Replace first dict in messages with a new system task. This is to distinguish between the two types of analysts
-                self.code_messages[0] = {"role": "system", "content": self.prompts.code_generator_system_gen}
+                self.message_manager.code_messages[0] = {"role": "system", "content": self.prompts.code_generator_system_gen}
             agent = 'Planner'
 
         elif expert == 'Research Specialist':
-            self.eval_messages.append({"role": "user", "content": self.prompts.theorist_system.format(utils.get_readable_date(),
+            self.message_manager.eval_messages.append({"role": "user", "content": self.prompts.theorist_system.format(utils.get_readable_date(),
                                                                                               None if self.df_id is None else df_columns,
                                                                                               aux_datasets_columns, 
-                                                                                              self.last_code, self.format_qa_pairs(self.qa_pairs), 
+                                                                                              self.message_manager.last_code, self.message_manager.format_qa_pairs(), 
                                                                                               question)
                                                                                               })
             agent = 'Theorist'
             self.output_manager.display_results(chain_id=self.chain_id, query={"expert":expert, "original_question": question, "unknown": query_unknown, "condition": query_condition, "requires_dataset": requires_dataset, "confidence": confidence, "intent_breakdown": intent_breakdown})
         else:
-            self.eval_messages.append({"role": "user", "content": self.prompts.theorist_system.format(utils.get_readable_date(), 
+            self.message_manager.eval_messages.append({"role": "user", "content": self.prompts.theorist_system.format(utils.get_readable_date(), 
                                                                                               None if self.df_id is None else df_columns,
                                                                                               aux_datasets_columns,   
-                                                                                              self.last_code, self.format_qa_pairs(self.qa_pairs), 
+                                                                                              self.message_manager.last_code, self.message_manager.format_qa_pairs(), 
                                                                                               question)
                                                                                               })
             agent = 'Theorist'
         
         if not analyst or (self.planning and  models.get_model_name("Code Generator")[0] not in self.reasoning_models): # If the analyst is not selected or the code generator model is not a reasoning model
-            task_eval,tool_response, full_response = self.task_eval(self.eval_messages, agent, image)
-            self.eval_messages.append({"role": "assistant", "content": full_response})
-            self.last_plan = task_eval
+            task_eval,tool_response, full_response = self.task_eval(self.message_manager.eval_messages, agent, image)
+            self.message_manager.eval_messages.append({"role": "assistant", "content": full_response})
+            self.message_manager.last_plan = task_eval
         else:
-            task_eval = self.last_plan # If exists it will be included in the messages under header CONTEXT:. This is to pass Theorist output to coder if planning is disabled, or model is a reasoning model
+            task_eval = self.message_manager.last_plan # If exists it will be included in the messages under header CONTEXT:. This is to pass Theorist output to coder if planning is disabled, or model is a reasoning model
             tool_response = []
-            self.last_plan = None
+            self.message_manager.last_plan = None
 
         # Remove the oldest conversation and all tool calls, and dataset, plan, and code from the messages list
-        self.messages_maintenace(self.eval_messages)
-        self.messages_content_maintenance(agent, self.eval_messages)
+        self.message_manager.messages_maintenace(self.message_manager.eval_messages)
+        self.message_manager.messages_content_maintenance(agent, self.message_manager.eval_messages, self.model_dict[models.get_model_name('Code Generator')[0]]['templ_formating'])
 
         plan = task_eval
 
@@ -761,33 +485,7 @@ class BambooAI:
         # If chain_id is provided and different from current, restore the content of that chain to serve as the starting point
         if chain_id is not None and str(chain_id) != str(self.chain_id):
             try:
-                if hasattr(self, 'interaction_store') and self.interaction_store is not None:
-                    restored_data = self.interaction_store.restore_interaction(
-                        thread_id=str(self.thread_id),
-                        chain_id=str(chain_id)
-                    )
-                    
-                    # Set the messages from restored data
-                    self.pre_eval_messages = restored_data['pre_eval_messages']
-                    self.select_analyst_messages = restored_data['select_analyst_messages']
-                    self.eval_messages = restored_data['eval_messages']
-                    self.df_inspector_messages = restored_data['df_inspector_messages']
-                    self.code_messages = restored_data['code_messages']
-                    self.plan_review_messages = restored_data['plan_review_messages']
-                    self.insight_messages = restored_data['insight_messages']
-
-                    # Restore code execution results
-                    self.code_exec_results = restored_data['code_exec_results']
-
-                    # Restore the last executed code
-                    self.last_code = restored_data['executed_code']
-
-                    # Restore the QA pairs
-                    self.qa_pairs = restored_data['qa_pairs']
-                    
-                    # Restore the tasks
-                    self.tasks = restored_data['tasks']
-                    
+                self.message_manager.restore_interaction(self.thread_id, chain_id)
             except storage_manager.StorageError as e:
                 self.output_manager.print_wrapper(f"Warning: Failed to restore chain {chain_id}: {e}")
         
@@ -805,7 +503,7 @@ class BambooAI:
                 self._process_question(question, image, user_code)
 
                 # Reset non-cumulative messages
-                self.reset_non_cumul_messages()
+                self.message_manager.reset_non_cumul_messages()
                 
                 if not is_web:
                     self.log_and_call_manager.consolidate_logs()
@@ -840,8 +538,8 @@ class BambooAI:
                 if not analyst:
                     self.output_manager.display_results(chain_id=self.chain_id, research=tool_response, answer=plan)
                     self.log_and_call_manager.print_summary_to_terminal(self.output_manager)
-                    self.store_interaction(google_search_results=tool_response, code_exec_results=self.code_exec_results, executed_code=self.last_code, qa_pairs=self.qa_pairs, tasks=self.tasks)
-                    self.tasks.append(question)
+                    self.message_manager.store_interaction(self.thread_id, self.chain_id, google_search_results=tool_response, code_exec_results=self.message_manager.code_exec_results, executed_code=self.message_manager.last_code, qa_pairs=self.message_manager.qa_pairs, tasks=self.message_manager.tasks)
+                    self.message_manager.tasks.append(question)
                     return
                 else:
                     if self.planning and models.get_model_name("Code Generator")[0] not in self.reasoning_models:
@@ -873,7 +571,7 @@ class BambooAI:
                 analyst, 
                 intent_breakdown, 
                 plan, 
-                self.code_messages, 
+                self.message_manager.code_messages, 
                 example_code, 
                 image=image, 
                 generated_datasets_path=generated_datasets_path
@@ -884,8 +582,8 @@ class BambooAI:
             if code is None or code == "":
                 self.output_manager.display_results(chain_id=self.chain_id, answer=llm_response)
                 self.log_and_call_manager.print_summary_to_terminal(self.output_manager)
-                self.store_interaction(google_search_results=tool_response, code_exec_results=self.code_exec_results, executed_code=self.last_code, qa_pairs=self.qa_pairs, tasks=self.tasks)
-                self.tasks.append(intent_breakdown)
+                self.message_manager.store_interaction(self.thread_id, self.chain_id, google_search_results=tool_response, code_exec_results=self.message_manager.code_exec_results, executed_code=self.message_manager.last_code, qa_pairs=self.message_manager.qa_pairs, tasks=self.message_manager.tasks)
+                self.message_manager.tasks.append(intent_breakdown)
                 return
         else:
             analyst = 'User'
@@ -897,7 +595,7 @@ class BambooAI:
             code_type = 'user'
 
         # Call the execute_code() method to execute the code and summarise the results
-        answer, results, code, plot_jsons, error_corrections = self.execute_code(analyst, code, plan, intent_breakdown, self.code_messages, self.execution_mode, code_type, generated_datasets_path)
+        answer, results, code, plot_jsons, error_corrections = self.execute_code(analyst, code, plan, intent_breakdown, self.message_manager.code_messages, self.execution_mode, code_type, generated_datasets_path)
 
         # Review and correct the plan for storage if there were code corrections or use the original plan
         reviewed_plan = None
@@ -946,16 +644,16 @@ class BambooAI:
         self.log_and_call_manager.print_summary_to_terminal(self.output_manager)
 
         # Append the QA pair to the list of QA pairs
-        self.append_qa_pair(intent_breakdown, results)
+        self.message_manager.append_qa_pair(intent_breakdown, results)
 
         # Append Task to the list of tasks
-        self.tasks.append(intent_breakdown)
+        self.message_manager.tasks.append(intent_breakdown)
 
         # Reset the retrieved data
         self.reset_retrieved_data()
 
         # Store the interaction
-        self.store_interaction(google_search_results=tool_response, plot_jsons=plot_jsons, code_exec_results=results, executed_code=code, qa_pairs=self.qa_pairs, tasks=self.tasks)
+        self.message_manager.store_interaction(self.thread_id, self.chain_id, google_search_results=tool_response, plot_jsons=plot_jsons, code_exec_results=results, executed_code=code, qa_pairs=self.message_manager.qa_pairs, tasks=self.message_manager.tasks)
             
     ######################
     ### Code Functions ###
@@ -989,7 +687,7 @@ class BambooAI:
             python_version=self.python_version,
             pandas_version=self.pandas_version,
             plotly_version=self.plotly_version,
-            previous_results=self.format_qa_pairs(self.qa_pairs),
+            previous_results=self.message_manager.format_qa_pairs(),
             example_code=example_code
         )
 
@@ -1003,7 +701,7 @@ class BambooAI:
         # Format messages if image is provided
         if image is not None:
             last_message = code_messages[-1]
-            formatted_message = self.format_image_message(
+            formatted_message = self.message_manager.format_image_message(
                 agent=agent,
                 message_content=last_message["content"],
                 image=image,
@@ -1040,7 +738,7 @@ class BambooAI:
 
         while error_corrections < self.MAX_ERROR_CORRECTIONS:
             # Remove the oldest conversation from the messages list
-            self.messages_maintenace(code_messages)
+            self.message_manager.messages_maintenace(code_messages)
 
             # Execute the code
             if code is not None:
@@ -1061,19 +759,18 @@ class BambooAI:
 
             # Remove examples from the messages list to minimize the number of tokens used
             code_messages = reg_ex._remove_examples(code_messages)
-            
+
         if self.webui:
             # If webui is enabled, we display the results of code execution using the output manager
             self.output_manager.display_results(chain_id=self.chain_id, code_exec_results=results)
 
-
         # Remove dataset, plan, and code from the messages list
-        self.messages_content_maintenance(agent, code_messages)
+        self.message_manager.messages_content_maintenance(agent, code_messages, self.model_dict[models.get_model_name('Code Generator')[0]]['templ_formating'])
 
         # Store the results in a class variable so it can be appended to the subsequent messages list
-        self.code_exec_results = results
+        self.message_manager.code_exec_results = results
         # Store the last generated code in a class variable so it can be appended to the subsequent messages list
-        self.last_code = code
+        self.message_manager.last_code = code
 
         plot_jsons = []
         for i, plot_data in enumerate(plot_images):
@@ -1128,15 +825,15 @@ class BambooAI:
         reasoning_effort = "low"
 
         # Initialize the messages list with a user message containing the task prompt
-        self.plan_review_messages = [{"role": "user", "content": self.prompts.reviewer_system.format(code, plan)}]
+        self.message_manager.plan_review_messages = [{"role": "user", "content": self.prompts.reviewer_system.format(code, plan)}]
 
         using_model,provider = models.get_model_name(agent)
 
         self.output_manager.display_tool_start(agent,using_model, chain_id=self.chain_id)
 
-        llm_response = self.llm_stream(self.prompts, self.log_and_call_manager, self.output_manager, self.plan_review_messages, agent=agent, chain_id=self.chain_id, reasoning_models=self.reasoning_models, reasoning_effort=reasoning_effort)
+        llm_response = self.llm_stream(self.prompts, self.log_and_call_manager, self.output_manager, self.message_manager.plan_review_messages, agent=agent, chain_id=self.chain_id, reasoning_models=self.reasoning_models, reasoning_effort=reasoning_effort)
 
-        self.plan_review_messages.append({"role": "assistant", "content": llm_response})
+        self.message_manager.plan_review_messages.append({"role": "assistant", "content": llm_response})
 
         reviewed_plan = reg_ex._extract_plan(llm_response)
         plan_vis = utils.generate_plan_graph(plan)
@@ -1161,17 +858,17 @@ class BambooAI:
 
         # Initialize the messages list with a user message containing the task prompt
         if code is not None:
-            self.insight_messages = [{"role": "user", "content": self.prompts.solution_summarizer_custom_code_system.format(code, results)}]
+            self.message_manager.insight_messages = [{"role": "user", "content": self.prompts.solution_summarizer_custom_code_system.format(code, results)}]
         else:
-            self.insight_messages = [{"role": "user", "content": self.prompts.solution_summarizer_system.format(intent_breakdown, plan, results)}]
+            self.message_manager.insight_messages = [{"role": "user", "content": self.prompts.solution_summarizer_system.format(intent_breakdown, plan, results)}]
         
         using_model,provider = models.get_model_name(agent)
 
         self.output_manager.display_tool_start(agent,using_model, chain_id=self.chain_id)
 
-        summary = self.llm_stream(self.prompts, self.log_and_call_manager, self.output_manager, self.insight_messages, agent=agent, chain_id=self.chain_id, reasoning_models=self.reasoning_models, reasoning_effort=reasoning_effort)
+        summary = self.llm_stream(self.prompts, self.log_and_call_manager, self.output_manager, self.message_manager.insight_messages, agent=agent, chain_id=self.chain_id, reasoning_models=self.reasoning_models, reasoning_effort=reasoning_effort)
 
-        self.insight_messages.append({"role": "assistant", "content": summary})
+        self.message_manager.insight_messages.append({"role": "assistant", "content": summary})
 
         self.output_manager.display_results(chain_id=self.chain_id, answer=summary)
